Fix tests: move location manager tests to locations.test.tsx
The location manager tests were still targeting AdminSettingsScreen.
Move them to a new locations.test.tsx pointing at AdminLocationsScreen,
and slim settings.test.tsx down to only cover what remains there.
